Cache Page import in global and move to django hook
