+++ newrelic/config.py
@@ -2995,7 +2995,7 @@ def _process_module_builtin_defaults():
     _process_module_definition("flask_restx.api", "newrelic.hooks.component_flask_rest", "instrument_flask_rest")
 
     _process_module_definition(
-        "wagtail.models.pages", "newrelic.hooks.framework_wagtail", "instrument_wagtail_models_pages"
+        "wagtail.models.pages", "newrelic.hooks.framework_django", "instrument_wagtail_models_pages"
     )
 
     _process_module_definition("graphql_server", "newrelic.hooks.component_graphqlserver", "instrument_graphqlserver")
